requirements.txt
discord.py==1.7.3
playsound==1.3.0
pyserial==3.5
python-dotenv==1.0.0
pyzmq==23.2.0
pydub==0.25.1